use test_name rather than __LINE__ for hidden test structs
When building tests from mutliple files together, the line number the test is
declared on is not guaranteed to be unique. If two tests happen to be on the
same line, one of them will not be run. You will then go crazy trying to figure
out why seemingly unrelated changes make the problem go away.