python bindings: don't try to index on dict_keys error with python3
This fixes a problem when using the bindings with python3 when passing
python dictionaries will result in the error:

TypeError: 'dict_keys' object does not support indexing

This is because the keys() method of a dictionary no longer returns an
object that is indexable. Instead of indexing, use the iterator's
'__next__' method to get the next object the appropiate number of times.