don't create web seed connections if the torrent is upload only
The fix in 9a63d4696ebec5743418181ef3411371d8e31b04 was imcomplete.
It turns out disconnect_if_redundant decrements the connection count
which unballances the count if we do end up disconnecting the peer.

This change avoids the problem by checking if the torrent is
upload only much sooner, before the connection object is even created.
We still do a redundancy check just-in-case, but it is not expected to
trigger.