Fix build on FreeBSD 11.1
Build fails on FreeBSD 11.1-p1 using cmake and base clang compiler.
I'm not sure what the correct convention is here for this iconv for
all BSD version, but certainly casting it to non-const and then
passing it in should be safe even if some versions of iconv take
a const pointer.