Make emplace_alert lock-free
The main benefit of this patch is that it de-couples the library's
internal threads from the user threads by making emplace_alert() not
longer mutually exclusive with pop_alerts() and wait_for_alert(), except
for a short period after a call to set_queue_size_limit() until it takes
effect on or before the next call to pop_alerts(). So it eliminates
contention and makes wait_for_alert() more deterministic.

It also makes emplace_alert() completely thread-safe and replaces the
coarse locking mechanism with very fine-grained locking so that many
threads can safely make progress through emplace_alert()

To allow concurrent access the alerts themselves are allocated on the
heap and pointers are stored on a ring buffer atomically.