Cleanup and fix race conditions
Fix race conditions when setting m_dispatch and m_notify by replacing
the reference counter with a shared lock which is really what it was but
now it's used to protect the callback functions as well. Since
set_notify_function() and set_dispatch_function() now must acquire an
exclusive lock that makes them much higher latency since now they wait
for the changes to take effect. It actually forces them to take effect
immediately but since that requires stopping all threads it takes much
longer. So any tests that stress these functions will take much longer
to complete. It should not be a problem with real world usage and it
makes the code much easier to read since it gets rid of most of the
relaxed loads.