Tweak Peer Turnover prioritisation...
... to prefer peers still waiting to download over the slowest actually downloading.

Peer Turnover only applies when most connection slots are in use, and the primary purpose is to free up the least beneficial connection slots in the hope that LibTorrent can find a faster download peer.

Previously a connection which has been made but which has not yet started downloading is preferred for disconnect over the slowest downloading connection.

The primary aim of this change is to prioritise connections that are waiting to start downloading over the slowest actually downloading on  the assumption that a connection that has already been made but is waiting to start downloading has the same chance of being faster than any new connection - and we should not lose the time already spent waiting by disconnecting and making a new connection that then waits. 

This is perhaps a bit simplistic, because it leaves LibTorrent open to keeping (potentially malicious) connections open that are (perhaps deliberately) never going to start downloading. To avoid this we will limit the above change to connections which have been waiting less than Peer_Turnover_Maximum_Wait and retain the current preference of disconnecting the waiting connection if it has been waiting longer than this, For this commit Peer_Turnover_Maximum_Wait is a hardcoded constant but this can be changed to a configurable setting in a future commit.