Improve superseeding performance
In some cases (smaller piece size, higher upload capacity) upload speed to peers of superseeded torrents may be limited when compared to standard seeding as the peer can only queue up for download data that we've told it we have - at present a max of 2 pieces.

This change increases the number of pieces we tell a peer we have based on our upload speed to them. We offer as many pieces as we expect them to download in superseed_scale_time seconds (defaults to 6s) at a time.