Turn `has_tracker_query_string` more efficient
In this implementation, i use a map to store the tracker arguments, allowing the lookup to be done in O(log N) time instead of O(N) time in the original loop. Additionally, i also removed the need to do a case-insensitive comparison using `string_equal_no_case`, which can also be a performance gain.