Allow loading .torrent files with 1 GB piece size
Previously the limit was 1073741823, however 1 GB = 1073741824 which is off by one (at least
for x64 systems).
The reason to allow 1 GB piece size is because this is the maximum size that could be created
by libtorrent. And it doesn't make sense if one can create it but cannot load it successfully.