Make anti-leech choker never increase score for dishonesty
The current anti-leech choker takes the maximum of client-reported have
and our own session uploaded size as the progress of the client.  This
may paradoxically inflate the score of dishonest clients if we've been
uploading to it for long enough, that our uploaded size exceeds 50% of
the total size.

This commit caps "given_size" to 50% of total_size, so that the above-
mentioned situation can never happen.